requirements.txt
selenium>=4.15,<5
requests>=2.31,<3
beautifulsoup4>=4.12,<5
python-dotenv>=1.0,<2
supabase>=2.4,<3
Pillow>=10.0,<12
